ogar_io/requirements.txt
flask==1.0.2
pygame==1.9.4
requests-2.20.1
